fix: bug arquivo que nao existe no bonus

+++ mandatory/src/parse/textures_obtaining.c
@@ -5,8 +5,8 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: luizedua <luizedua@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/01/13 18:09:27 by lucperei          #+#    #+#             */
-/*   Updated: 2024/01/18 15:02:02 by luizedua         ###   ########.fr       */
+/*   Created: 2024/01/17 20:45:42 by lucperei          #+#    #+#             */
+/*   Updated: 2024/01/23 12:56:16 by luizedua         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,13 +23,13 @@ void	cal_wall_dimension(t_game *game, int coord[2], int invert, int *offset)
 	else
 		offset[0] = calc_horizontal_disp(game->rays[coord[0]]);
 	dist_to_top = coord[1] + (floor(height) / 2) - ((float)WIN_H / 2);
-	offset[1] = dist_to_top * (64.0 / floor(height));
+	offset[1] = dist_to_top * ((float)64 / floor(height));
 }
 
 static int	get_texture_color(t_image texture, int offset[2])
 {
 	return (*(unsigned int *)(texture.i_add + (offset[1] * texture.size_line + \
-			offset[0] * (texture.bpp / 8))));
+		offset[0] * (texture.bpp / 8))));
 }
 
 int	get_texture(t_game *game, t_image texture, int coord[2], int invert)
